requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.16
torch-cluster>=1.6.0
torch-spline-conv>=1.2.1
networkx>=3.0
numpy>=1.23.0
scipy>=1.9.0
pandas>=1.5.0
scikit-learn>=1.1.0
transformers>=4.25.0
spacy>=3.4.0
PyPDF2>=3.0.0
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.6.0
plotly>=5.10.0
pygraphviz>=1.9
optuna>=3.0.0
ray[tune]>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
black>=22.10.0
flake8>=5.0.0
pytest>=7.2.0
pytest-cov>=4.0.0
ipython>=8.5.0
jupyter>=1.0.0